auth-aoth/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyJWT
python-dotenv
bcrypt
